+++ packages/common/src/ide/inMemoryTextDocument/InMemoryTextDocumentEdit.test.ts
@@ -1,41 +1,34 @@
 import * as assert from "node:assert";
-import { URI } from "vscode-uri";
 import { Range } from "../..";
-import { InMemoryTextDocument } from "./InMemoryTextDocument";
-import { performEdits } from "./performEdits";
-
-const uri = URI.parse("cursorless-dummy://dummy/untitled");
-const languageId = "plaintext";
+import { createDocument } from "./createDocument";
 
 const text = "hello\nworld";
 
-suite("performEdits", () => {
+suite("InMemoryTextDocument edit", () => {
   test("change", async () => {
-    const document = new InMemoryTextDocument(uri, languageId, text);
-    await performEdits(document, [
-      { range: new Range(0, 0, 0, 5), text: "goodbye" },
-    ]);
+    const document = createDocument(text);
+    await document.edit([{ range: new Range(0, 0, 0, 5), text: "goodbye" }]);
 
     assert.equal(document.text, "goodbye\nworld");
   });
 
   test("remove", async () => {
-    const document = new InMemoryTextDocument(uri, languageId, text);
-    await performEdits(document, [{ range: new Range(0, 0, 1, 0), text: "" }]);
+    const document = createDocument(text);
+    await document.edit([{ range: new Range(0, 0, 1, 0), text: "" }]);
 
     assert.equal(document.text, "world");
   });
 
   test("insert", async () => {
-    const document = new InMemoryTextDocument(uri, languageId, text);
-    await performEdits(document, [{ range: new Range(0, 5, 0, 5), text: "!" }]);
+    const document = createDocument(text);
+    await document.edit([{ range: new Range(0, 5, 0, 5), text: "!" }]);
 
     assert.equal(document.text, "hello!\nworld");
   });
 
   test("multiple", async () => {
-    const document = new InMemoryTextDocument(uri, languageId, text);
-    await performEdits(document, [
+    const document = createDocument(text);
+    await document.edit([
       { range: new Range(0, 5, 0, 5), text: "!" },
       { range: new Range(1, 0, 1, 1), text: "" },
       { range: new Range(0, 0, 0, 5), text: "goodbye" },
@@ -45,8 +38,8 @@ suite("performEdits", () => {
   });
 
   test("remove overlapping", async () => {
-    const document = new InMemoryTextDocument(uri, languageId, text);
-    const changes = await performEdits(document, [
+    const document = createDocument(text);
+    const changes = await document.edit([
       { range: new Range(0, 0, 0, 4), text: "" },
       { range: new Range(0, 1, 1, 1), text: "" },
     ]);
@@ -57,10 +50,10 @@ suite("performEdits", () => {
   });
 
   test("change overlapping", async () => {
-    const document = new InMemoryTextDocument(uri, languageId, text);
+    const document = createDocument(text);
 
     try {
-      await performEdits(document, [
+      await document.edit([
         { range: new Range(0, 0, 0, 4), text: " " },
         { range: new Range(0, 1, 1, 1), text: "" },
       ]);
